% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5api.R
\name{getPostCarnage}
\alias{getPostCarnage}
\title{Get post-carnage report for given mode and match}
\usage{
getPostCarnage(mode = "", match_id = "", key = "")
}
\description{
Get post-carnage report for given mode and match
}

