% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5api.R
\name{getProfileEmblem}
\alias{getProfileEmblem}
\title{Get spartan profile emblem image}
\usage{
getProfileEmblem(url = url_profile, player, size = 256, key = "")
}
\description{
Get spartan profile emblem image
}

