% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5api.R
\name{getProfileSpartan}
\alias{getProfileSpartan}
\title{Get spartan profile spartan image}
\usage{
getProfileSpartan(player, size = 256, crop = "full", key = "")
}
\description{
Get spartan profile spartan image
}

