% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5api.R
\name{getRecentMatches}
\alias{getRecentMatches}
\title{Get a player's most recent matches}
\usage{
getRecentMatches(url = url_player_stats, player = "",
  modes = "arena,campaign,custom,warzone", start = 0, count = 25,
  key = "")
}
\description{
Get a player's most recent matches
}

