% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5api.R
\name{getServiceRecord}
\alias{getServiceRecord}
\title{Get spartan service record for a given mode}
\usage{
getServiceRecord(mode = "", players = "", key = "")
}
\description{
Get spartan service record for a given mode
}

