% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{h5_Leaderboard}
\alias{h5_Leaderboard}
\title{Get CSR leaderboards for playlists}
\usage{
h5_Leaderboard(playlist = "", season = "", count = 200, key = "")
}
\description{
Get CSR leaderboards for playlists
}
