% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{h5_PlayerGameVariant}
\alias{h5_PlayerGameVariant}
\title{Get information about a player-created game variant}
\usage{
h5_PlayerGameVariant(player = "", variant = "", key = "")
}
\description{
Get information about a player-created game variant
}
