% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{h5_PlayerGameVariants}
\alias{h5_PlayerGameVariants}
\title{Get a list of game variants created by a player}
\usage{
h5_PlayerGameVariants(player = "", start = 0, count = 20, key = "")
}
\description{
Get a list of game variants created by a player
}
