% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{h5_PlayerMapVariant}
\alias{h5_PlayerMapVariant}
\title{Get information about a player-created map variant}
\usage{
h5_PlayerMapVariant(player = "", variant = "", key = "")
}
\description{
Get information about a player-created map variant
}
