% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{h5_PlayerMapVariants}
\alias{h5_PlayerMapVariants}
\title{Get a list of map variants created by a player}
\usage{
h5_PlayerMapVariants(player = "", start = 0, count = 20, key = "")
}
\description{
Get a list of map variants created by a player
}
