% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{h5_PostCarnage}
\alias{h5_PostCarnage}
\title{Get post-carnage report for given mode and match}
\usage{
h5_PostCarnage(mode = "", match_id = "", key = "")
}
\description{
Get post-carnage report for given mode and match
}
