% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{h5_ProfileSpartan}
\alias{h5_ProfileSpartan}
\title{Get spartan profile spartan image}
\usage{
h5_ProfileSpartan(player, size = 256, crop = "full", key = "")
}
\description{
Get spartan profile spartan image
}
