% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{h5_RecentMatches}
\alias{h5_RecentMatches}
\title{Get a player's most recent matches}
\usage{
h5_RecentMatches(player = "", modes = c("arena", "campaign", "custom",
  "warzone"), start = 0, count = 25, key = "")
}
\description{
Get a player's most recent matches
}
