% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{h5_ServiceRecord}
\alias{h5_ServiceRecord}
\title{Get spartan service record for a given mode}
\usage{
h5_ServiceRecord(mode = "", players = "", key = "", season = "")
}
\description{
Get spartan service record for a given mode
}
