% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{hw2_GameObjectCategories}
\alias{hw2_GameObjectCategories}
\title{Get a list of game object categories}
\usage{
hw2_GameObjectCategories(startAt = 0, key = "")
}
\description{
Get a list of game object categories
}
