% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{hw2_PlayerCampaignProgress}
\alias{hw2_PlayerCampaignProgress}
\title{Get a player's campaign progress}
\usage{
hw2_PlayerCampaignProgress(player = "", key = "")
}
\description{
Get a player's campaign progress
}
