% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{hw2_PlayerMatchHistory}
\alias{hw2_PlayerMatchHistory}
\title{Get a player's match history
matchType can be 'custom' or 'matchmaking', or empty for both}
\usage{
hw2_PlayerMatchHistory(player = "", matchType = "", start = 0,
  count = 25, key = "")
}
\description{
Get a player's match history
matchType can be 'custom' or 'matchmaking', or empty for both
}
