% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{hw2_PlayerPlaylistRatings}
\alias{hw2_PlayerPlaylistRatings}
\title{Get a player's playlist ratings}
\usage{
hw2_PlayerPlaylistRatings(playlist = "", players = "", key = "")
}
\description{
Get a player's playlist ratings
}
