% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haloR.R
\name{hw2_PlayerSeasonStats}
\alias{hw2_PlayerSeasonStats}
\title{Get a player's season stats summary}
\usage{
hw2_PlayerSeasonStats(player = "", season = "current", key = "")
}
\description{
Get a player's season stats summary
}
